Mixin Utility class

class that holds a few util methods used across all classes in the this R package

Methods

Public methods

MixinUtilities$create_empty_dataframe()

MixinUtilities$list_of_vectors_to_dataframe()

MixinUtilities$apply_func_to_df()

MixinUtilities$clone()

Method create_empty_dataframe()

Usage

MixinUtilities$create_empty_dataframe(col_names)

Method list_of_vectors_to_dataframe()

Usage

MixinUtilities$list_of_vectors_to_dataframe(lht_list)

Method apply_func_to_df()

Usage

MixinUtilities$apply_func_to_df(data, func_space)

Method clone()

The objects of this class are cloneable with this method.

Usage

MixinUtilities$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.